Renamed IMediaType and implementations to IHttpRequestBodyConverter.
