+++ src/app/SharpRaven/Data/IHttpRequestBodyConverter.cs
@@ -2,21 +2,21 @@
 
 // Copyright (c) 2014 The Sentry Team and individual contributors.
 // All rights reserved.
-//
+// 
 // Redistribution and use in source and binary forms, with or without modification, are permitted
 // provided that the following conditions are met:
-//
+// 
 //     1. Redistributions of source code must retain the above copyright notice, this list of
 //        conditions and the following disclaimer.
-//
+// 
 //     2. Redistributions in binary form must reproduce the above copyright notice, this list of
 //        conditions and the following disclaimer in the documentation and/or other materials
 //        provided with the distribution.
-//
+// 
 //     3. Neither the name of the Sentry nor the names of its contributors may be used to
 //        endorse or promote products derived from this software without specific prior written
 //        permission.
-//
+// 
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
 // IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
 // FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
@@ -28,24 +28,38 @@
 
 #endregion
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace SharpRaven.Data
 {
-    public class MultiPartFormMediaType : FormMediaType
+    /// <summary>
+    /// HTTP media type interface for converting the HTTP body of a request to a structured type.
+    /// </summary>
+    public interface IHttpRequestBodyConverter
     {
-        public override bool Matches(string mediaType)
-        {
-            if (String.IsNullOrEmpty(mediaType))
-            {
-                return false;
-            }
+        /// <summary>
+        /// Checks whether the specified <paramref name="contentType"/> can be converted by this
+        /// <see cref="IHttpRequestBodyConverter"/> implementation or not.
+        /// </summary>
+        /// <param name="contentType">The media type to match.</param>
+        /// <returns>
+        /// Returns <c>true</c> if the <see cref="IHttpRequestBodyConverter"/> implementation can convert
+        /// the specified <paramref name="contentType"/> cref="contentType"/>; otherwise <c>false</c>.
+        /// </returns>
+        bool Matches(string contentType);
 
-            var mimeType = mediaType.Split(';').First();
 
-            return mimeType.Equals("multipart/form-data", StringComparison.OrdinalIgnoreCase);
-        }
+        /// <summary>
+        /// Tries to convert the HTTP request body of the specified <paramref name="httpContext"/> to
+        /// a structured type.
+        /// </summary>
+        /// <param name="httpContext">The HTTP context containing the request body to convert.</param>
+        /// <param name="converted">
+        /// The converted, structured type for the specified <paramref name="httpContext"/>'s request
+        /// body or <c>null</c> if the <paramref name="httpContext"/> is null, or the somehow conversion
+        /// fails.
+        /// </param>
+        /// <returns>
+        /// <c>true</c> if the conversion succeeds; otherwise <c>false</c>.
+        /// </returns>
+        bool TryConvert(dynamic httpContext, out object converted);
     }
 }
