+++ src/app/SharpRaven/SharpRaven.csproj
@@ -81,15 +81,15 @@
     <Compile Include="Data\BreadcrumbLevel.cs" />
     <Compile Include="Data\BreadcrumbType.cs" />
     <Compile Include="Data\Breadcrumb.cs" />
-    <Compile Include="Data\DefaultMediaType.cs" />
+    <Compile Include="Data\DefaultHttpRequestBodyConverter.cs" />
     <Compile Include="Data\ExceptionData.cs" />
-    <Compile Include="Data\MultiPartFormMediaType.cs" />
-    <Compile Include="Data\IMediaType.cs" />
+    <Compile Include="Data\MultiPartFormHttpRequestBodyConverter.cs" />
+    <Compile Include="Data\IHttpRequestBodyConverter.cs" />
     <Compile Include="Data\ISentryRequestFactory.cs" />
     <Compile Include="Data\ISentryUserFactory.cs" />
     <Compile Include="Data\ISentryRequest.cs" />
-    <Compile Include="Data\FormMediaType.cs" />
-    <Compile Include="Data\JsonMediaType.cs" />
+    <Compile Include="Data\FormHttpRequestBodyConverter.cs" />
+    <Compile Include="Data\JsonHttpRequestBodyConverter.cs" />
     <Compile Include="Data\JsonPacketFactory.cs" />
     <Compile Include="Data\ErrorLevel.cs" />
     <Compile Include="Data\ExceptionFrame.cs" />
